+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## [Unreleased]
 
+## [0.0.2] - 2024-07-15
+### Added
+- Support for 'basic' module functions and operators
+- Code formatting support with customizable settings
+### Improved
+- Syntax highlighting for 'basic' module functions
+
 ## [0.0.1] - 2024-07-01
 ### Added
 - Initial release